Create Updated list and dictionary,
there are a few missing parts, but from what I thought of testing most of list is working. Its just missing the part where two variables on a list would add their values. and the methods from the dictionary. It should display whole lists. One thing, I set it up as a programs this way i thought it would be easier to test. It should be no problem adding them to the class